+++ custom_components/stiebel_dhe_connect/const.py
@@ -10,6 +10,7 @@ DEFAULT_NAME = "DHE Connect"
 DEFAULT_PORT = 8443
 
 PLATFORMS = [
+    Platform.BINARY_SENSOR,
     Platform.CLIMATE,
     Platform.MEDIA_PLAYER,
     Platform.WEATHER,